giorno 4

+++ app_blogging/src/main/java/it/epicode/app_blogging/blog_autore/AutoreService.java
@@ -1,5 +1,6 @@
 package it.epicode.app_blogging.blog_autore;
 
+import it.epicode.app_blogging.email.EmailService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,8 +13,13 @@ public class AutoreService {
     @Autowired
     private AutoreRepository repository;
 
+    @Autowired
+    private EmailService emailService;
+
     public Autore create(Autore autore) {
-        return repository.save(autore);
+        Autore savedAutore = repository.save(autore);
+        emailService.sendConfermaAutore(autore.getEmail(), "Benvenuto!", "Ciao " + autore.getNome() + ", benvenuto nel nostro sistema!");
+        return savedAutore;
     }
 
     public List<Autore> getAll() {
@@ -25,7 +31,16 @@ public class AutoreService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        Optional<Autore> autore = repository.findById(id);
+
+        if (autore.isPresent()) {
+            // Chiamare il servizio EmailService per inviare una notifica prima di eliminare
+            emailService.sendConfermaAutore(autore.get().getEmail(), "Eliminazione Account", "Ciao " + autore.get().getNome() + ", il tuo account è stato eliminato.");
+
+            // Elimina l'autore
+            repository.deleteById(id);
+        }
     }
+
 }
